pyproject.toml
[project]
name = "docops-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "patriciorequena"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.109.0)",
    "uvicorn[standard] (>=0.27.0)",
    "pydantic (>=2.5.0)",
    "pydantic-settings (>=2.1.0)",
    "sqlalchemy (>=2.0.25)",
    "asyncpg (>=0.29.0)",
    "alembic (>=1.13.0)",
    "minio (>=7.2.0)",
    "redis (>=5.0.1)",
    "celery (>=5.3.6)",
    "pymupdf (>=1.23.0)",
    "unstructured[pdf] (>=0.12.0)",
    "tabula-py (>=1.0.8)",
    "python-multipart (>=0.0.6)",
    "google-generativeai (>=0.4.0)",
    "langgraph (>=0.0.20)",
    "chromadb (>=0.4.22)",
    "langchain (>=0.1.0)",
    "httpx (>=0.26.0)",
    "streamlit (>=1.54.0,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    { include = "src", from = "." }
]

[tool.poetry.group.dev.dependencies]
langsmith = "^0.7.3"
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"

